+++ crates/ika-sui-client/src/lib.rs
@@ -49,6 +49,7 @@ use sui_types::{
 use tokio::sync::OnceCell;
 use tracing::{debug, error, info, warn};
 
+pub mod dwallet_signer;
 pub mod ika_dwallet_transactions;
 #[cfg(feature = "protocol-commands")]
 pub mod ika_protocol_transactions;
